# openapi.yaml
openapi: 3.0.3
info:
  title: RAG AI Platform API
  version: "2.0.0"
  description: |
    Unified API contract for the full platform:
    - RAG application service (chat/session/tools/upload endpoints)
    - Backend data service (team/investments/sectors/documents/scrape endpoints)

    Notes:
    - Backend endpoints (except `/auth/token`) require `Authorization: Bearer <token>`.
    - RAG `/api/*` endpoints require gateway bearer auth only when `ENABLE_GATEWAY_AUTH=true`.
  contact:
    name: Son Nguyen
    url: https://sonnguyenhoang.com
    email: hoangson091104@gmail.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Local RAG API
  - url: http://localhost:3456
    description: Local backend API
  - url: http://rag-app:5000
    description: Internal Docker network (RAG service)
  - url: http://backend:3456
    description: Internal Docker network (backend service)
...
components:
  securitySchemes:
    backendBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Required by backend endpoints except `/auth/token`.
    ragGatewayBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Required only when `ENABLE_GATEWAY_AUTH=true` on the RAG service.
  parameters:
    XRequestID:
      name: X-Request-ID
      in: header
      required: false
      schema:
        type: string
      description: Optional client-supplied request ID for traceability.
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Unauthorized
        request_id:
          type: string
          example: 9dd87863-c3e6-443f-b509-033f3fe4de95
...
    SystemInfoResponse:
      type: object
      required:
        - success
        - app_name
        - api_version
        - environment
        - llm_model
        - embedding_model
        - rerank_model
        - max_query_chars
        - available_strategies
        - backend_tools
        - request_id
      properties:
        success:
          type: boolean
          example: true
        app_name:
          type: string
        api_version:
          type: string
          example: v1
        environment:
          type: string
        llm_model:
          type: string
          example: llama2
        embedding_model:
          type: string
          example: sentence-transformers/all-MiniLM-L6-v2
        rerank_model:
          type: string
          example: cross-encoder/ms-marco-MiniLM-L-6-v2
        max_query_chars:
          type: integer
          example: 4000
        available_strategies:
          type: array
          items:
            type: string
            enum:
              - semantic
              - hybrid
              - multi_query
              - decomposed
        backend_tools:
          type: array
          items:
            type: string
          example:
            - team_profile
            - investment_profile
            - sector_profile
        request_id:
          type: string
    ToolsResponse:
      type: object
      required:
        - success
        - tools
        - request_id
      properties:
        success:
          type: boolean
          example: true
        tools:
          type: array
          items:
            type: string
          example:
            - team_profile
            - team_insights
            - investment_profile
            - investment_insights
            - sector_profile
            - consultations
            - scrape_page
        request_id:
          type: string
...
    ChatResult:
      type: object
      required:
        - query
        - response
        - strategy
        - num_documents
        - sources
        - api_data_keys
        - api_chain_trace
        - metadata
      properties:
        query:
          type: string
        response:
          type: string
        strategy:
          type: string
          enum:
            - semantic
            - hybrid
            - multi_query
            - decomposed
        num_documents:
          type: integer
          example: 5
        sources:
          type: array
          items:
            $ref: "#/components/schemas/SourceItem"
        api_data_keys:
          type: array
          items:
            type: string
        api_chain_trace:
          type: array
          items:
            $ref: "#/components/schemas/AgenticTraceItem"
        metadata:
          type: object
          additionalProperties: true
          example:
            cache_hit: false
            latency_ms: 1023.5
            timestamp: 2026-02-06T23:35:00Z
            api_chain_calls: 3
    ChatRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 1
          example: Summarize current portfolio opportunities and risks.
        strategy:
          type: string
          enum:
            - semantic
            - hybrid
            - multi_query
            - decomposed
          default: hybrid
        session_id:
          type: string
          nullable: true
    ChatResponse:
      type: object
      required:
        - success
        - session_id
        - result
        - request_id
      properties:
        success:
          type: boolean
          example: true
        session_id:
          type: string
          nullable: true
        result:
          $ref: "#/components/schemas/ChatResult"
        request_id:
          type: string
    OpenAIChatMessage:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          example: user
        content:
          type: string
    OpenAIChatCompletionRequest:
      type: object
      required:
        - messages
      properties:
        model:
          type: string
          nullable: true
          example: llama2
        strategy:
          type: string
          enum:
            - semantic
            - hybrid
            - multi_query
            - decomposed
          default: hybrid
        session_id:
          type: string
          nullable: true
        messages:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OpenAIChatMessage"
    OpenAIChatCompletionChoice:
      type: object
      required:
        - index
        - message
        - finish_reason
      properties:
        index:
          type: integer
          example: 0
        message:
          $ref: "#/components/schemas/OpenAIChatMessage"
        finish_reason:
          type: string
          example: stop
    OpenAIUsage:
      type: object
      required:
        - prompt_tokens
        - completion_tokens
        - total_tokens
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
    OpenAIChatCompletionResponse:
      type: object
      required:
        - id
        - object
        - created
        - model
        - choices
        - usage
        - metadata
        - request_id
      properties:
        id:
          type: string
          example: chatcmpl-1234567890abcdef12345678
        object:
          type: string
          example: chat.completion
        created:
          type: integer
          example: 1738890000
        model:
          type: string
          example: llama2
        choices:
          type: array
          items:
            $ref: "#/components/schemas/OpenAIChatCompletionChoice"
        usage:
          $ref: "#/components/schemas/OpenAIUsage"
        metadata:
          type: object
          additionalProperties: true
        request_id:
          type: string
    SessionMessage:
      type: object
      required:
        - timestamp
        - query
        - response
        - strategy
      properties:
        timestamp:
          type: string
          format: date-time
        query:
          type: string
        response:
          type: string
        strategy:
          type: string
          enum:
            - semantic
            - hybrid
            - multi_query
            - decomposed
    SessionData:
      type: object
      required:
        - created_at
        - messages
      properties:
        created_at:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/SessionMessage"
    SessionCreatedResponse:
      type: object
      required:
        - success
        - session_id
        - created_at
        - messages
        - request_id
      properties:
        success:
          type: boolean
          example: true
        session_id:
          type: string
        created_at:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/SessionMessage"
        request_id:
          type: string
    SessionEnvelopeResponse:
      type: object
      required:
        - success
        - session_id
        - session
        - request_id
      properties:
        success:
          type: boolean
          example: true
        session_id:
          type: string
        session:
          $ref: "#/components/schemas/SessionData"
        request_id:
          type: string
...
    SessionListResponse:
      type: object
      required:
        - success
        - sessions
        - total
        - request_id
      properties:
        success:
          type: boolean
          example: true
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/SessionListItem"
        total:
          type: integer
          example: 1
        request_id:
          type: string
    DeleteSessionResponse:
      type: object
      required:
        - success
        - request_id
      properties:
        success:
          type: boolean
          example: true
        request_id:
          type: string
    UploadResponse:
      type: object
      required:
        - success
        - filename
        - stored_as
        - added_chunks
        - message
        - request_id
      properties:
        success:
          type: boolean
          example: true
        filename:
          type: string
          example: market-brief.md
        stored_as:
          type: string
          example: a1b2c3d4_market-brief.md
        added_chunks:
          type: integer
          example: 12
        message:
          type: string
          example: File uploaded and indexed successfully.
        request_id:
          type: string
...
    StrategiesResponse:
      type: object
      required:
        - success
        - strategies
        - request_id
      properties:
        success:
          type: boolean
          example: true
        strategies:
          type: array
          items:
            $ref: "#/components/schemas/StrategyItem"
        request_id:
          type: string
paths:
...
  /api/system/info:
    get:
      tags:
        - RAG System
      summary: RAG runtime configuration and capability snapshot
      operationId: ragSystemInfo
      security:
        - ragGatewayBearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/XRequestID"
      responses:
        "200":
          description: System info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemInfoResponse"
        "401":
          description: Unauthorized (when gateway auth is enabled)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/tools:
    get:
      tags:
        - RAG System
      summary: Available backend tools for agentic API chaining
      operationId: ragTools
      security:
        - ragGatewayBearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/XRequestID"
      responses:
        "200":
          description: Tool registry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolsResponse"
        "401":
          description: Unauthorized (when gateway auth is enabled)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/chat:
    post:
      tags:
        - RAG Chat
      summary: Query the RAG assistant
      operationId: ragChat
      security:
        - ragGatewayBearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/XRequestID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Chat response with source evidence and agentic trace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (when gateway auth is enabled)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/chat/completions:
    post:
      tags:
        - RAG Chat
      summary: OpenAI-compatible chat completion endpoint
      operationId: ragChatCompletions
      security:
        - ragGatewayBearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/XRequestID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenAIChatCompletionRequest"
      responses:
        "200":
          description: OpenAI-compatible completion payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAIChatCompletionResponse"
        "400":
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (when gateway auth is enabled)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/session:
    post:
      tags:
        - RAG Sessions
      summary: Create a new chat session
      operationId: ragCreateSession
      security:
        - ragGatewayBearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/XRequestID"
      responses:
        "200":
          description: Created session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionCreatedResponse"
        "401":
          description: Unauthorized (when gateway auth is enabled)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/session/{session_id}:
    get:
      tags:
        - RAG Sessions
      summary: Fetch session content by ID
      operationId: ragGetSession
      security:
        - ragGatewayBearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/XRequestID"
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionEnvelopeResponse"
        "401":
          description: Unauthorized (when gateway auth is enabled)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - RAG Sessions
      summary: Delete a session
      operationId: ragDeleteSession
      security:
        - ragGatewayBearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/XRequestID"
        - name: session_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Session deleted (idempotent behavior)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteSessionResponse"
        "401":
          description: Unauthorized (when gateway auth is enabled)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/sessions:
    get:
      tags:
        - RAG Sessions
      summary: List known sessions
      operationId: ragListSessions
      security:
        - ragGatewayBearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/XRequestID"
      responses:
        "200":
          description: Session list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionListResponse"
        "401":
          description: Unauthorized (when gateway auth is enabled)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/upload:
    post:
      tags:
        - RAG Knowledge
      summary: Upload and index a knowledge document
      operationId: ragUploadDocument
      security:
        - ragGatewayBearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/XRequestID"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: File uploaded and indexed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
        "400":
          description: Invalid file, unsupported extension, or missing payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (when gateway auth is enabled)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/strategies:
    get:
      tags:
        - RAG Knowledge
      summary: List retrieval strategies supported by the RAG engine
      operationId: ragStrategies
      security:
        - ragGatewayBearerAuth: []
        - {}
      parameters:
        - $ref: "#/components/parameters/XRequestID"
      responses:
        "200":
          description: Available retrieval strategies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategiesResponse"
        "401":
          description: Unauthorized (when gateway auth is enabled)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...